[PHP] Extract database row reading from MySQLDumpProducer
